Browser issue figured out / resolved.
In the Chrome developer tools, cursor was a semi-transparent circle; apparently this represents a "finger" on mobile devices?  Obviously, a "hover" state or mouseover does not function the same way as on laptop / desktop computers.  Added two versions of laptops under "emulated devices" and everything works properly (for larger screens / non-handheld) when they are selected.