Finished test code; YAY...IT WORKS!!!
I know it is working properly, and effectively "tests" what I was lookin for.  Initially, I wired the event handler (onclick) to function test() which stacked entries in the console.  This way, I call the reload() function and use location.reload, then kick off function test() thereafter.  Not only is it working, but the output no longer "stacks" in console output.